--- a/R/dials-package.R
+++ b/R/dials-package.R
@@ -34,11 +34,11 @@
 ## usethis namespace: start
 #' @importFrom lifecycle deprecated
 #' @importFrom pillar type_sum
-#' @importFrom purrr map_lgl map2_dfc map_chr map map2 map_dfc map_dbl
+#' @importFrom purrr map_lgl map2_dfc map_chr map map2 map_dfc map_dbl map_int
 #' @importFrom DiceDesign dmaxDesign lhsDesign
 #' @importFrom rlang quos eval_tidy quo_get_expr is_string enquo quo_is_null
 #' @importFrom rlang caller_env check_dots_empty
-#' @importFrom tibble as_tibble is_tibble tibble
+#' @importFrom tibble as_tibble is_tibble tibble new_tibble
 #' @importFrom scales log2_trans is.trans log10_trans
 #' @importFrom utils installed.packages globalVariables
 #' @importFrom glue glue glue_collapse